+++ py/fiberassign/stucksky.py
@@ -102,9 +102,8 @@ def stuck_on_sky(hw, tiles):
         stuck_sky[tileid][loc] = bool_good_sky
     '''
     from fiberassign.utils import Logger
-    from fiberassign.hardware import FIBER_STATE_STUCK, FIBER_STATE_BROKEN
-
-    from fiberassign.targets import xy2radec
+    from fiberassign.hardware import (FIBER_STATE_STUCK, FIBER_STATE_BROKEN,
+                                      xy2radec)
 
     log = Logger.get()
