+++ py/fiberassign/targets.py
@@ -32,6 +32,7 @@ from desitarget.sv3.sv3_targetmask import desi_mask as sv3_mask
 from desitarget.targets import main_cmx_or_sv
 
 from .utils import Logger, Timer
+from .hardware import radec2xy
 
 from ._internal import (TARGET_TYPE_SCIENCE, TARGET_TYPE_SKY,
                         TARGET_TYPE_STANDARD, TARGET_TYPE_SAFE,
@@ -937,54 +938,6 @@ def load_target_file(tgs, tfile, survey=None, typeforce=None, typecol=None,
     tm.report("Read target file {}".format(tfile))
 
     return survey
-
-def radec2xy(hw, tile_ra, tile_dec, tile_obstime, tile_obstheta, tile_obsha,
-             ra, dec, use_cs5, threads=0):
-    #xy = hw.radec2xy_multi(
-    #    tile_ra, tile_dec, tile_obstheta, ra, dec, use_cs5, threads=0
-    #)
-    #x = np.array([x for x,y in xy])
-    #y = np.array([y for x,y in xy])
-    from astropy.time import Time
-    from desimeter.fiberassign import fiberassign_radec2xy_cs5, fiberassign_radec2xy_flat
-    # Note that MJD is only used for precession, so no need for
-    # high precision.
-    t = Time(tile_obstime, format='isot')
-    mjd = t.mjd
-
-    # Don't pass adc[12]: Let desimeter use its pm-alike routines
-    if use_cs5:
-        x, y = fiberassign_radec2xy_cs5(ra, dec, tile_ra, tile_dec, mjd,
-                                        tile_obsha, tile_obstheta)
-    else:
-        x, y = fiberassign_radec2xy_flat(ra, dec, tile_ra, tile_dec, mjd,
-                                         tile_obsha, tile_obstheta)
-    return x,y
-
-def xy2radec(hw, tile_ra, tile_dec, tile_obstime, tile_obstheta, tile_obsha,
-             x, y, use_cs5, threads=0):
-    # radec = hw.xy2radec_multi(
-    #     tile_ra, tile_dec, tile_obstheta, x, y, use_cs5, threads
-    #     )
-    # ra  = np.array([r for r,d in radec])
-    # dec = np.array([d for r,d in radec])
-    from desimeter.fiberassign import fiberassign_cs5_xy2radec, fiberassign_flat_xy2radec
-    from astropy.time import Time
-    t = Time(tile_obstime, format='isot')
-    mjd = t.mjd
-    if use_cs5:
-        ra,dec = fiberassign_cs5_xy2radec(x, y, tile_ra, tile_dec, mjd,
-                                          tile_obsha, tile_obstheta)
-    else:
-        ra,dec = fiberassign_flat_xy2radec(x, y, tile_ra, tile_dec, mjd,
-                                           tile_obsha, tile_obstheta)
-    return ra,dec
-
-def xy2cs5(x, y):
-    # There's a change in terminology between the focal-plane team and
-    # the outside world here...
-    from desimeter.transform.pos2ptl import ptl2flat
-    return flat2ptl(x, y)
 
 def targets_in_tiles(hw, tgs, tiles):
     '''
